Move resources
