+++ health-entity-web/src/app/components/resources/allergy-intolerance/create-allergy-intolerance/create-allergy-intolerance.component.html
