+++ health-entity-web/src/app/components/resources/allergy-intolerance/read-allergy-intolerance/read-allergy-intolerance.component.spec.ts
