+++ health-entity-web/src/app/components/resources/allergy-intolerance/read-allergy-intolerance/read-allergy-intolerance.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { AllergyIntolerance } from '../../../models/allergy-intolerance';
+import { AllergyIntolerance } from '../../../../models/allergy-intolerance';
 import { Reaction } from 'app/models/reaction';
 import { OptionsList } from 'app/models/options-lists';
