+++ health-entity-web/src/app/components/resources/condition/create-condition/create-condition.component.html
