+++ health-entity-web/src/app/components/resources/condition/read-condition/read-condition.component.ts
