+++ health-entity-web/src/app/components/resources/diagnostic-report/create-diagnostic-report/create-diagnostic-report.component.scss
