+++ health-entity-web/src/app/components/resources/diagnostic-report/read-diagnostic-report/read-diagnostic-report.component.html
