+++ health-entity-web/src/app/components/resources/observation/create-observation/create-observation.component.html
