+++ health-entity-web/src/app/components/resources/observation/read-observation/read-observation.component.scss
