+++ health-entity-web/src/app/components/resources/observation/read-observation/read-observation.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Observation } from '../../../models/observation';
+import { Observation } from '../../../../models/observation';
 import { OptionsList } from 'app/models/options-lists';
 
 @Component({
