+++ health-entity-web/src/app/components/resources/procedure/create-procedure/create-procedure.component.scss
