+++ health-entity-web/src/app/components/resources/procedure/read-procedure/read-procedure.component.scss
