+++ health-entity-web/src/app/models/diagnostic-report.ts
@@ -1,5 +1,6 @@
 import { Identification } from './identification';
 import { Observation } from './observation';
+
 export class DiagnosticReport {
   constructor(
     public identifier: number,
