+++ health-entity-web/src/app/services/security/auth-guard.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Router, CanActivate, ActivatedRouteSnapshot } from '@angular/router';
-import { AuthService } from './auth-service';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root'
